requirements-in.txt
# Image processing
numpy==1.24.3
scipy==1.10.1
pillow==9.5.0
opencv-python==4.7.0.72

# Device connection
playwright==1.50.0
asyncio

# Utils
rich==13.3.5
tqdm==4.65.0
pyyaml==6.0
inflection==0.5.1
prettytable==2.2.1
pydantic>=2.4
onepush==1.3.0
filelock==3.17.0
pytz==2024.2
tzlocal==5.2
lxml==4.9.1
beautifulsoup4==4.13.4
redis==6.4.0
orjson==3.11.2
psutil==7.0.0
portalocker==3.2.0

# OCR
pponnxcr==2.0

# Webui
pywebio==1.8.3
starlette==0.14.2
uvicorn[standard]==0.17.6
aiofiles

# Game resources
srcmap==2.3.20240701

# For dev
# pip-tools
pynput==1.7.6
